Refactor : Session 패키지 생성 후 관련 클래스 이동
