+++ src/main/java/http/session/SessionHandler.java
@@ -1,4 +1,4 @@
-package http;
+package http.session;
 
 import model.User;
